Refactor app to use a test client for certbot in dev and test
**ISSUE**
Certbot is intended to run in production environments. Configuring
Certbot to run successfully in a local development or test
environment would be complicated and fragile.  It also requires
sudo priveleges that block tests and development servers unless
local sudoers is modified.

**SOLUTION**
Default to use the Certbot::V2::TestClient in development and
test environments instead of Certbot::V2::Client, which is
only used in production.
